utils: move to klplib package
