+++ klpbuild/klplib/ksrc.py
@@ -15,7 +15,7 @@ from pathlib import PurePath
 import requests
 from natsort import natsorted
 
-from klpbuild import utils
+from klpbuild.klplib import utils
 from klpbuild.config import Config
 from klpbuild.klplib.codestream import Codestream
 from klpbuild.klplib.ibs import IBS
