+++ klpbuild/plugins/setup.py
@@ -9,7 +9,7 @@ from pathlib import Path
 
 from natsort import natsorted
 
-from klpbuild import utils
+from klpbuild.klplib import utils
 from klpbuild.config import Config
 from klpbuild.klplib.ksrc import GitHelper
