+++ tests/test_extract.py
@@ -8,7 +8,7 @@ import logging
 
 from klpbuild.extractor import Extractor
 from klpbuild.plugins.setup import Setup
-from klpbuild import utils
+from klpbuild.klplib import utils
 
 
 def test_detect_file_without_ftrace_support(caplog):
