+++ tests/test_templ.py
@@ -7,7 +7,7 @@ import inspect
 
 from klpbuild.extractor import Extractor
 from klpbuild.plugins.setup import Setup
-from klpbuild import utils
+from klpbuild.klplib import utils
 from tests.utils import get_file_content
